Rewrite TracedData to take advantage of Python's built in views
