+++ core_data_modules/traced_data.py
@@ -1,11 +1,10 @@
+from collections import Mapping, KeysView, ValuesView, ItemsView, Iterator
+
 from deprecation import deprecated
 
 import six
 
 from core_data_modules.util.sha_utils import SHAUtils
-from core_data_modules.traced_data_views.traced_data_items_view import _TracedDataItemsView
-from core_data_modules.traced_data_views.traced_data_keys_view import _TracedDataKeysView
-from core_data_modules.traced_data_views.traced_data_values_view import _TracedDataValuesView
 
 
 class Metadata(object):
@@ -18,19 +17,25 @@ class Metadata(object):
         return self.user == other.user and self.source == other.source and self.timestamp == other.timestamp
 
 
-class TracedData(object):
+class TracedData(Mapping):
     def __init__(self, data, metadata, _prev=None):
         self._prev = _prev
         self._data = data
         self._sha = self._sha_with_prev(data, "" if _prev is None else _prev._sha)
         self._metadata = metadata
 
+    def append(self, new_data, new_metadata):
+        self._prev = TracedData(self._data, self._metadata, self._prev)
+        self._data = new_data
+        self._sha = self._sha_with_prev(self._data, self._prev._sha)
+        self._metadata = new_metadata
+
     @staticmethod
     def _sha_with_prev(data, prev_sha):
         return SHAUtils.sha_dict({"data": data, "prev_sha": prev_sha})
 
-    def __len__(self):
-        return sum(1 for _ in six.viewkeys(self))
+    def __getitem__(self, key):
+        return self.get(key)
 
     def get(self, key, default=None):
         if key in self._data:
@@ -40,24 +45,6 @@ class TracedData(object):
         else:
             return default
 
-    def __getitem__(self, key):
-        return self.get(key)
-
-    def append(self, new_data, new_metadata):
-        self._prev = TracedData(self._data, self._metadata, self._prev)
-        self._data = new_data
-        self._sha = self._sha_with_prev(self._data, self._prev._sha)
-        self._metadata = new_metadata
-
-    @deprecated
-    def has_key(self, key):
-        if self._data.has_key(key):
-            return True
-        elif self._prev is not None:
-            return self._prev.has_key(key)
-        else:
-            return False
-
     def __contains__(self, key):
         if key in self._data:
             return True
@@ -66,64 +53,55 @@ class TracedData(object):
         else:
             return False
 
-    if six.PY3:
-        def items(self):
-            return _TracedDataItemsView(self)
+    def __iter__(self):
+        return _TracedDataKeysIterator(self)
 
-        def keys(self):
-            return _TracedDataKeysView(self)
-
-        def values(self):
-            return _TracedDataValuesView(self)
+    def __len__(self):
+        return sum(1 for _ in self)
 
     if six.PY2:
-        def items(self):
-            if self._prev is None:
-                return self._data.items()
-            else:
-                prev_items = dict(self._prev.items())
-                for (key, value) in self._data.items():
-                    prev_items[key] = value
-                return prev_items.items()
-
-        def iteritems(self):
-            return iter(self.viewitems())
-
-        def viewitems(self):
-            return _TracedDataItemsView(self)
+        @deprecated
+        def has_key(self, key):
+            return key in self
 
         def keys(self):
-            if self._prev is None:
-                return self._data.keys()
-            else:
-                prev_items = dict(self._prev.items())
-                for (key, value) in self._data.items():
-                    prev_items[key] = value
-                return prev_items.keys()
-
-        def iterkeys(self):
-            return iter(self.viewkeys())
-
-        def viewkeys(self):
-            return _TracedDataKeysView(self)
+            return list(self)
 
         def values(self):
-            if self._prev is None:
-                return self._data.values()
-            else:
-                prev_items = dict(self._prev.items())
-                for (key, value) in self._data.items():
-                    prev_items[key] = value
-                return prev_items.values()
+            return [self[key] for key in self]
+
+        def items(self):
+            return [(key, self[key]) for key in self]
+
+        def iterkeys(self):
+            return iter(self)
 
         def itervalues(self):
-            return iter(self.viewvalues())
+            for key in self:
+                yield self[key]
+
+        def iteritems(self):
+            for key in self:
+                yield (key, self[key])
+
+        def viewkeys(self):
+            return KeysView(self)
 
         def viewvalues(self):
-            return _TracedDataValuesView(self)
+            return ValuesView(self)
 
-    def __iter__(self):
-        return six.iterkeys(self)
+        def viewitems(self):
+            return ItemsView(self)
+
+    if six.PY3:
+        def keys(self):
+            return KeysView(self)
+
+        def values(self):
+            return ValuesView(self)
+
+        def items(self):
+            return ItemsView(self)
 
     def __eq__(self, other):
         return self._data == other._data and self._metadata == other._metadata and \
@@ -138,3 +116,38 @@ class TracedData(object):
         if key in self._data:
             history.append({"sha": self._sha, "value": self._data[key]})
         return history
+
+
+# noinspection PyProtectedMember
+class _TracedDataKeysIterator(Iterator):
+    def __init__(self, traced_data):
+        self.traced_data = traced_data
+        self.next_keys = iter(traced_data._data.keys())
+        self.seen_keys = set()
+
+    def __iter__(self):
+        return self
+
+    def _next_item(self):
+        while True:
+            # Search for a new key in the iterator for the current TracedData instance.
+            try:
+                while True:
+                    key = next(self.next_keys)
+                    if key not in self.seen_keys:
+                        self.seen_keys.add(key)
+                        return key
+            except StopIteration:
+                # We ran out of keys which we haven't yet returned. Try the prev TracedData.
+                self.traced_data = self.traced_data._prev
+                if self.traced_data is None:
+                    raise StopIteration()
+                self.next_keys = iter(self.traced_data._data.keys())
+
+    if six.PY2:
+        def next(self):
+            return self._next_item()
+
+    if six.PY3:
+        def __next__(self):
+            return self._next_item()
